messengyr v0.0.1 Messengyr.Web.Router.Helpers

Module with named helpers generated from Messengyr.Web.Router.

Functions

chat_path(conn_or_endpoint, action)

chat_path(conn_or_endpoint, action, params)

chat_url(conn_or_endpoint, action)

chat_url(conn_or_endpoint, action, params)

message_path(conn_or_endpoint, action, id)

message_path(conn_or_endpoint, action, id, params)

message_url(conn_or_endpoint, action, id)

message_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

room_path(conn_or_endpoint, action)

room_path(conn_or_endpoint, action, params)

room_path(conn_or_endpoint, action, id, params)

room_url(conn_or_endpoint, action)

room_url(conn_or_endpoint, action, params)

room_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action, id)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action, id)

user_url(conn_or_endpoint, action, id, params)